refactor: extract GitHub integration into collavre_github engine
- Create new collavre_github engine with isolated namespace
- Move models: GithubAccount -> CollavreGithub::Account, GithubRepositoryLink -> CollavreGithub::RepositoryLink
- Move controllers: webhooks, accounts, auth, integrations
- Move services: Client, WebhookProvisioner, PullRequestProcessor, PullRequestAnalyzer
- Move views: integration modal
- Update routes to use engine mounting
- Update model aliases to use CollavreGithub namespace
- Remove GitHub-related code from collavre engine and host app
- Add i18n support (en, ko) for the new engine

This follows the same pattern as collavre_notion and collavre_slack engines,
making GitHub integration optional and maintaining clear dependency boundaries.
