+++ config/routes.rb
@@ -38,20 +38,6 @@ Rails.application.routes.draw do
     resource :settings, only: [ :update ]
   end
 
-  # Creative integrations (GitHub/Notion)
-  resources :creatives, only: [] do
-    resource :github_integration, only: [ :show, :update, :destroy ], module: :creatives
-  end
-
-  # GitHub account and webhooks
-  namespace :github do
-    resource :account, only: [ :show ] do
-      get :organizations
-      get :repositories
-    end
-    post :webhook, to: "webhooks#create"
-  end
-
   # MCP OAuth Discovery
   get "/.well-known/oauth-protected-resource", to: "mcp/discovery#oauth_protected_resource"
   get "/.well-known/oauth-authorization-server", to: "mcp/discovery#oauth_authorization_server"
