+++ engines/collavre_github/Rakefile
@@ -1,0 +1,8 @@
+require "bundler/setup"
+
+APP_RAKEFILE = File.expand_path("../test/dummy/Rakefile", __FILE__)
+load "rails/tasks/engine.rake"
+
+load "rails/tasks/statistics.rake"
+
+require "bundler/gem_tasks"
