+++ engines/collavre_github/app/models/collavre_github/application_record.rb
@@ -1,0 +1,5 @@
+module CollavreGithub
+  class ApplicationRecord < ActiveRecord::Base
+    self.abstract_class = true
+  end
+end
