+++ engines/collavre_github/config/locales/en.yml
@@ -1,0 +1,34 @@
+en:
+  collavre_github:
+    integration:
+      label: "GitHub"
+      description: "Connect repositories and receive webhook events"
+      title: "Configure Github integration"
+      saved: "Github integration saved successfully."
+      login_required: "Sign in with your Github account to start the integration."
+      missing_creative: "No Creative selected for integration."
+      webhook_url_label: "Webhook URL"
+      webhook_secret_label: "Webhook secret"
+      existing_message: "You're already connected to the repositories below."
+      existing_intro: "Connected repositories:"
+      delete_confirm: "Do you want to remove the Github integration?"
+      delete_success: "Github integration removed successfully."
+      delete_error: "Failed to remove the Github integration."
+      delete_button: "Remove integration"
+      delete_select_warning: "Select repositories to delete."
+      connect: "Sign in with your Github account to start linking."
+      login_button: "Sign in with Github"
+      choose_org: "Select the organization that owns the repositories."
+      choose_repo: "Select repositories to link. You can choose multiple."
+      summary: "The following repositories will be linked to this Creative:"
+      webhook_instructions: "Configure each repository with the webhook details below."
+      summary_empty: "No repositories selected."
+      prompt_title: "Review the Gemini analysis prompt and edit it if needed."
+      prompt_help: "Use the placeholders below to inject runtime details as needed:"
+    auth:
+      login_first: "Please log in first to connect your GitHub account"
+      connected: "GitHub account connected successfully"
+    errors:
+      not_connected: "GitHub account not connected"
+      forbidden: "You don't have permission to perform this action"
+      not_found: "Repository not found"
