+++ engines/collavre_github/config/locales/ko.yml
@@ -1,0 +1,34 @@
+ko:
+  collavre_github:
+    integration:
+      label: "GitHub"
+      description: "저장소를 연결하고 웹훅 이벤트를 받습니다"
+      title: "Github 연동 설정"
+      saved: "Github 연동이 저장되었습니다."
+      login_required: "Github 계정으로 로그인하여 연동을 시작하세요."
+      missing_creative: "연동할 Creative가 선택되지 않았습니다."
+      webhook_url_label: "Webhook URL"
+      webhook_secret_label: "Webhook 시크릿"
+      existing_message: "이미 아래 저장소와 연결되어 있습니다."
+      existing_intro: "연결된 저장소:"
+      delete_confirm: "Github 연동을 삭제하시겠습니까?"
+      delete_success: "Github 연동이 삭제되었습니다."
+      delete_error: "Github 연동 삭제에 실패했습니다."
+      delete_button: "연동 삭제"
+      delete_select_warning: "삭제할 Repository를 선택하세요."
+      connect: "Github 계정으로 로그인하여 연동을 시작하세요."
+      login_button: "Github으로 로그인"
+      choose_org: "저장소가 속한 조직을 선택하세요."
+      choose_repo: "연동할 저장소를 선택하세요. 여러 개 선택 가능합니다."
+      summary: "다음 저장소가 이 Creative에 연동됩니다:"
+      webhook_instructions: "각 저장소에 아래 웹훅 정보를 설정하세요."
+      summary_empty: "선택된 저장소가 없습니다."
+      prompt_title: "Gemini 분석 프롬프트를 검토하고 필요시 수정하세요."
+      prompt_help: "다음 플레이스홀더를 사용하여 런타임 정보를 주입할 수 있습니다:"
+    auth:
+      login_first: "GitHub 계정을 연결하려면 먼저 로그인하세요"
+      connected: "GitHub 계정이 연결되었습니다"
+    errors:
+      not_connected: "GitHub 계정이 연결되지 않았습니다"
+      forbidden: "이 작업을 수행할 권한이 없습니다"
+      not_found: "저장소를 찾을 수 없습니다"
