+++ engines/collavre_github/lib/collavre_github.rb
@@ -1,0 +1,5 @@
+require "collavre_github/version"
+require "collavre_github/engine"
+
+module CollavreGithub
+end
