+++ engines/collavre_github/lib/collavre_github/version.rb
@@ -1,0 +1,3 @@
+module CollavreGithub
+  VERSION = "0.1.0"
+end
