+++ engines/collavre_github/test/test_helper.rb
@@ -1,0 +1,4 @@
+# Configure Rails Environment
+ENV["RAILS_ENV"] = "test"
+
+require_relative "../../../test/test_helper"
